feat: [Assumed Role] Edit mode - change auth type toggle
Closes MPT-15383
